pyproject.toml
[project]
name = "law-news-analyzer"
version = "1.0.0"
description = "AI 기반 법률 분쟁 사건 자동 발굴 시스템"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0,<6.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    "django-filter>=23.5",
    "celery>=5.3",
    "redis>=5.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "openai>=1.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
    "gunicorn>=21.2",
    "drf-spectacular>=0.27",
    "google-genai>=1.64.0",
    "apscheduler>=3.11.2",
]

[tool.uv]
package = false
